refactor: 기본 정보 입력 단계를 TestBasicInfoStep으로 분리
