+++ src/features/test-create/ui/index.ts
@@ -2,3 +2,4 @@ export { FunnelLayout } from './FunnelLayout';
 export { TestCreateFunnel } from './TestCreateFunnel';
 export { CategorySelectSheet } from './CategorySelectSheet';
 export { TestRegisterStep } from './TestRegisterStep';
+export { TestBasicInfoStep } from './TestBasicInfoStep';
